+++ gaphor/SysML/diagramitems.py
@@ -1,2 +1,3 @@
+from gaphor.SysML.allocations import *
 from gaphor.SysML.blocks import *
 from gaphor.SysML.requirements import *